src/tests/requirements.txt
flake8
freezegun
isort
pkgversion
tox
vcrpy
vcrpy-unittest
pytest
pytest-cov
pytest-django
model-mommy